Split packages into pkg/{builder,sources,util}
