+++ pkg/source/sources.go
@@ -1,4 +1,4 @@
-package main
+package source
 
 import (
 	"encoding/json"
@@ -7,14 +7,12 @@ import (
 	"log"
 	"os"
 	"path/filepath"
-	"regexp"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/docker/docker/pkg/jsonmessage"
 	"github.com/fsouza/go-dockerclient"
-	"github.com/scraperwiki/git-prep-directory"
+	git "github.com/scraperwiki/git-prep-directory"
 )
 
 type ImageSource interface {
@@ -239,18 +237,4 @@ func PullProgressCopier(target io.Writer) (io.WriteCloser, <-chan error) {
 		}
 	}()
 	return wrappedWriter, errorC
-}
-
-func ParseRegistryImage(fullName string) (registry, repository string) {
-	var (
-		DotBeforeSlash = regexp.MustCompile("^[^/]+\\.[^/]+/")
-	)
-
-	if DotBeforeSlash.MatchString(fullName) {
-		parts := strings.SplitN(fullName, "/", 2)
-		return parts[0], parts[1]
-	}
-
-	// No registry specified
-	return "", fullName
 }
